Updating handicap service to be more accurate whith how handicaps currently work, and building base handicaps for players without an existing handicap for the league.

+++ src/main/java/space/knightdev/dghandicap/service/HandicapService.java
@@ -11,5 +11,5 @@ public interface HandicapService {
 
     PlayerHandicap updatePlayerHandicap(final PlayerHandicap player, final Double score, final boolean manualAdjust);
 
-    Double calculateHandicapRoundScore(final Player player, final CoursePlayerRound round);
+    Integer calculateHandicapRoundScore(final Player player, final CoursePlayerRound round);
 }
